Move compare.js to compare.mjs
So it can be imported in files executed via bin/sucrase-node.
