+++ root/static/scripts/common/utility/compare.mjs
