+++ root/static/scripts/relationship-editor/utility/updateRelationships.js
@@ -16,7 +16,7 @@ import {
 } from 'weight-balanced-tree/update';
 
 import {compare} from '../../common/i18n.js';
-import {compareStrings} from '../../common/utility/compare.js';
+import {compareStrings} from '../../common/utility/compare.mjs';
 import setMapDefault from '../../common/utility/setMapDefault.js';
 import type {
   RelationshipEditorStateT,
